+++ game/source/motor/motor_animation.cpp
@@ -2,7 +2,7 @@
 
 #include "cseries/cseries.hpp"
 
-bool __cdecl motor_animation_resource_busy(long mover_index, e_motor_resource resource)
+bool __cdecl motor_animation_resource_busy(int32 mover_index, e_motor_resource resource)
 {
 	return INVOKE(0x00B8FE10, motor_animation_resource_busy, mover_index, resource);
 }
